% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{collect_data}
\alias{collect_data}
\title{Collecting data from web}
\usage{
collect_data()
}
\description{
This function allows you pull data once and format it in a user friendly way
}
\examples{
collect_data()
}
