% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homicides_by_state.R
\name{homicides_by_state}
\alias{homicides_by_state}
\title{Collecting data from web}
\usage{
homicides_by_state()
}
\description{
This function allows you pull data once and format it in a user friendly way
}
\examples{
homicides_by_state()
}
